\name{DBF}
\alias{DBF}
\title{DBF test statistic}
\description{Returns the DBF test statistic. This function is intended for internal use.}
\usage{DBF(G, H)}
\arguments{
  \item{G}{Gower's centered inner product matrix arising from the distance matrix}
  \item{H}{The centered matrix encoding group membership of each sampling unit}
}
\value{The scalar-valued DBF test statistic.}
\author{Christopher Minas}
