\name{DBFTest-package}
\alias{DBFTest-package}
\alias{DBFTest}
\docType{package}
\title{DBF test of no difference between groups}
\description{The DBFTest package implements the DBF test of no difference between groups. 
}
\details{
\tabular{ll}{
Package: \tab DBFTest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-07\cr
License: \tab GPL-2\cr
}
Given sampling units from different groups, and a suitable distance measure applied to measurements observed on the sampling units, the null hypothesis that there is no difference between
groups is tested.
}
\author{Christopher Minas}
\references{
Minas, C., Waddell, S. J. and Montana, G. (2011). Distance-based differential analysis of gene curves. Bioinformatics

Minas, C. and Montana, G. (2012). Distance-based analysis of variance: approximate inference. Submitted to Statistical Analysis and Data Mining. arxiv: http://arxiv.org/abs/1205.2417  
}